[telepathy-mission-control] Run systemd service after lipstick. JB#54665
Add lipstick.service to the "After" line in the systemd service file.
Apply the change to the systemd service file patch.

This is required because otherwise SMS sending fails as
mission-control-5 is started too early related to lipstick. The service
seems to get into a some sort of limbo as it cannot connect to the
necessary components and refuses to send messages _until_ an inbound SMS
is recived.

Signed-off-by: Jussi Laakkonen <jussi.laakkonen@jolla.com>